Create EoT8mainfile.tex
This is the main LaTeX file to which all of the chapter files link. It shouldn't really contain too much of substance itself.

The reason for the EoT8 naming convention (the 8, specifically) is because this is probably the 8th iteration of this project. It is a bit hard to tell because some of the iterations were very very short lived (an example where I tried to do this from the homotopy type theory foundational system springs to mind).