Add all of the tex/etc files to the repository.
Added all of the files in my EoT folder to this repository. Was not entirely sure if it was necessary to add the aux files, but thought I might as well. Same goes for the pdf files, and so on. 

The reason for the 'EoT8' naming convention is because this is probably the 8th iteration this project has gone through. I say 'probably' because some of the iterations were very very short lived (the homotopy type theory iteration springs to mind most readily).