Finishing 2A.1 and starting 2A.2
I finished off the section on ordered pairs and products with a discussion of n-ary products (my footnotes specifically might need revision later on). I also started the section on relations, omitting the proof that domains and images exist for now so that I can make sure I do not actually call im(R) the 'range' of R. This will be something I will constantly need to check later on, as well, but right now I just need to make sure the original definition is correct.