Added to 2A.2, added an example to 2A.1
I added a few definitions to 2A.2, such as defining composition of relations and so on. I'm hoping that by defining this in general for relations, first, it will be easier to prove things algebraically for functions later one, without have to dig around in the set theoretics. 

I also added in the general membership relation as an example in 2A.1, because it seemed like it should really follow on from the < relation on omega. 

Finally, I added in the existence proof for domains and images that was omitted in the last edit.